requirements.txt
python-dotenv 
pymongo 
motor
cloudinary
fastapi
uvicorn
pydantic
bson
python-multipart

